+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+## 1.9.0
+  * Adds the Placeholder Feed Item Report to the list of available reports [#42](https://github.com/singer-io/tap-adwords/pull/42)
+
+## 1.8.2
+  * In the pre-transform hook, only coerce values to null if their trimmed value is exactly '--' [#41](https://github.com/singer-io/tap-adwords/pull/41)
+
+## 1.8.0
+  * Adds the Placeholder Report to the list of available reports [#40](https://github.com/singer-io/tap-adwords/pull/40)
+
+## 1.7.2
+  * Upgrades the `googleads` SDK and uses API version v201806 [#37](https://github.com/singer-io/tap-adwords/pull/37)
+  * Migrated from suds to zeep to match version of googleads SDK being used [#38](https://github.com/singer-io/tap-adwords/pull/38)
+
+## 1.6.0
+  * Adds support for bookmarking within an attribution window to prevent a hung sync [#36](https://github.com/singer-io/tap-adwords/pull/36)
+
+## 1.5.0
+  * Adds the Call Metrics Call Details, Display Keyword Performance, Display Topics Performance, Account Performance, and Video Performance Reports [#34](https://github.com/singer-io/tap-adwords/pull/34)
+
+## 1.4.0
+  * Adds the "Shopping Performance Report" to the list of available reports [#25](https://github.com/singer-io/tap-adwords/pull/25)
+
 ## 1.3.3
   * Allows the downloading of Reports to happen as a stream to improve memory consumption [#28](https://github.com/singer-io/tap-adwords/pull/28)